--- web/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "web",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build:css": "postcss static/css/styles.css -o static/css/output.css"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "devDependencies": {
-    "autoprefixer": "^10.4.20",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.12"
-  },
-  "dependencies": {
-    "postcss-cli": "^11.0.0"
-  }
-}
