+++ web/requirements.txt
@@ -1,26 +1,8 @@
-django==5.1.1
-async-generator==1.10
-attrs==21.4.0
-beautifulsoup4==4.10.0
-beautifultable==1.0.1
-certifi==2021.10.8
-cffi==1.15.0
-charset-normalizer==2.0.12
-cryptography==36.0.1
-h11==0.13.0
-idna==3.3
-outcome==1.1.0
-pycparser==2.21
-pyOpenSSL==22.0.0
-PySocks==1.7.1
-requests==2.27.1
-selenium==4.1.2
-sniffio==1.2.0
-sortedcontainers==2.4.0
-soupsieve==2.3.1
-trio==0.20.0
-trio-websocket==0.9.2
-urllib3==1.26.8
-wcwidth==0.2.5
-wsproto==1.1.0
-langchain-ollama==0.2.0
+streamlit 
+langchain 
+langchain_ollama
+selenium
+beautifulsoup4
+lxml 
+html5lib
+python-dotenv
