+++ web/sample.env
@@ -1,0 +1,1 @@
+CHROME_DRIVER_PATH = "./chromedriver.exe"
