--- web/scraper/scrape.py
@@ -1,57 +1,0 @@
-from selenium import webdriver
-from selenium.webdriver.chrome.service import Service
-from selenium.webdriver.chrome.options import Options
-from bs4 import BeautifulSoup
-import os
-
-# Path to your chromedriver executable
-CHROME_DRIVER_PATH = os.path.join(os.getcwd(), './web/chromedriver.exe')
-
-def scrape_website(website):
-    print("Starting Chrome WebDriver...")
-    
-    # Set up Chrome options
-    chrome_options = Options()
-    chrome_options.add_argument("--headless")  # Run in headless mode
-    chrome_options.add_argument("--no-sandbox")
-    chrome_options.add_argument("--disable-dev-shm-usage")
-
-    # Initialize WebDriver
-    service = Service(CHROME_DRIVER_PATH)
-    driver = webdriver.Chrome(service=service, options=chrome_options)
-    
-    try:
-        driver.get(website)
-        print("Navigated to:", website)
-
-        # Scraping the page content
-        html = driver.page_source
-        return html
-    finally:
-        driver.quit()  # Ensure the WebDriver is closed
-
-def extract_body_content(html_content):
-    soup = BeautifulSoup(html_content, "html.parser")
-    body_content = soup.body
-    if body_content:
-        return str(body_content)
-    return ""
-
-def clean_body_content(body_content):
-    soup = BeautifulSoup(body_content, "html.parser")
-
-    for script_or_style in soup(["script", "style"]):
-        script_or_style.extract()
-
-    # Get text or further process the content
-    cleaned_content = soup.get_text(separator="\n")
-    cleaned_content = "\n".join(
-        line.strip() for line in cleaned_content.splitlines() if line.strip()
-    )
-
-    return cleaned_content
-
-def split_dom_content(dom_content, max_length=6000):
-    return [
-        dom_content[i : i + max_length] for i in range(0, len(dom_content), max_length)
-    ]
