--- web/scraper/templates/list.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Scraped Websites</title>
-    {% load static %}
-    <link href="{% static 'css/output.css' %}" rel="stylesheet">
-</head>
-
-<body class="bg-gray-100 text-gray-800 font-sans antialiased">
-    <div class="container mx-auto p-8">
-        <h1 class="text-3xl font-bold mb-6 text-center text-blue-600">List of Scraped Websites</h1>
-        <div class="overflow-x-auto">
-            <table class="min-w-full bg-white shadow-md rounded-lg">
-                <thead class="bg-blue-500 text-white uppercase text-sm leading-normal">
-                    <tr>
-                        <th class="py-3 px-6 text-left">Alias</th>
-                        <th class="py-3 px-6 text-left">URL</th>
-                        <th class="py-3 px-6 text-left">Title</th>
-                        <th class="py-3 px-6 text-left">Scraped At</th>
-                        <th class="border px-4 py-2">Download JSON</th>
-                    </tr>
-                </thead>
-                <tbody class="text-sm font-light">
-                    {% for data in scraped_data %}
-                    <tr class="hover:bg-blue-100 transition-colors duration-300">
-                        <td class="py-3 px-6">{{ data.alias }}</td>
-                        <td class="py-3 px-6 truncate">{{ data.url }}</td>
-                        <td class="py-3 px-6">{{ data.title }}</td>
-                        <td class="py-3 px-6">{{ data.scraped_at }}</td>
-                        <td class="border px-4 py-2">
-                            <a href="{% url 'download_json' data.alias %}" class="text-blue-500 hover:underline">Download</a>
-                        </td>
-                    </tr>
-                    {% endfor %}
-                </tbody>
-            </table>
-        </div>
-        <br>
-        <a class="px-6 py-3 bg-teal-500 text-white rounded-lg shadow-lg hover:bg-teal-400 transition duration-300 ease-in-out transform hover:scale-110 hover:shadow-teal-500/50 focus:outline-none" href="{% url 'scrape_website' %}">Back to Scrape Page</a>
-    </div>
-</body>
-
-</html>
