--- web/scraper/templates/scrape.html
@@ -1,41 +1,0 @@
-<!-- scrape.html -->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Web Scrape</title>
-    {% load static %}
-    <link href="{% static 'css/output.css' %}" rel="stylesheet">
-</head>
-<body class="bg-gray-900 text-gray-100 min-h-screen flex items-center justify-center">
-    <div class="max-w-md w-full p-8 bg-gray-800 rounded-lg shadow-lg space-y-6 animate-fade-in">
-        <h1 class="text-4xl font-bold text-center text-teal-400">Web Scrape Tool</h1>
-
-        <form method="POST" class="space-y-4">
-            {% csrf_token %}
-
-            <div>
-                <label for="url" class="block text-sm font-medium text-gray-300">Website URL</label>
-                <input type="text" name="url" id="url" required 
-                    class="w-full px-4 py-2 mt-1 bg-gray-700 text-white rounded-lg border border-teal-500 focus:outline-none focus:ring-2 focus:ring-teal-400 transition duration-300 ease-in-out transform hover:scale-105">
-            </div>
-            
-            <div>
-                <label for="parse_description" class="block text-sm font-medium text-gray-300">Parse Description</label>
-                <textarea name="parse_description" id="parse_description" required 
-                    class="w-full px-4 py-2 mt-1 bg-gray-700 text-white rounded-lg border border-teal-500 focus:outline-none focus:ring-2 focus:ring-teal-400 transition duration-300 ease-in-out transform hover:scale-105"></textarea>
-            </div>
-            
-            <div class="flex justify-center">
-                <button type="submit" 
-                    class="px-6 py-3 bg-teal-500 text-white rounded-lg shadow-lg hover:bg-teal-400 transition duration-300 ease-in-out transform hover:scale-110 hover:shadow-teal-500/50 focus:outline-none">
-                    Scrape
-                </button>
-            </div>
-        </form>
-        <br>
-        <a class="px-6 py-3 bg-teal-500 text-white rounded-lg shadow-lg hover:bg-teal-400 transition duration-300 ease-in-out transform hover:scale-110 hover:shadow-teal-500/50 focus:outline-none" href="{% url 'list_scraped_websites' %}">View Scraped Websites</a>
-    </div>
-</body>
-</html>
