--- web/tailwind.config.js
@@ -1,11 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    './scraper/templates/**/*.html', 
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
-
