--- web/webscrape/urls.py
@@ -1,28 +1,0 @@
-"""
-URL configuration for webscrape project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/5.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-from django.contrib import admin
-from django.urls import path
-from scraper import views
-
-urlpatterns = [
- 
-    path('admin/', admin.site.urls),
-    path('', views.home, name='home'),
-    path('scrape/', views.scrape_website, name='scrape_website'),
-    path('list/', views.list_scraped_websites, name='list_scraped_websites'),
-    path('download_json/<str:alias>/', views.download_json, name='download_json'),
-]
